[Feature] Ability to set/change data after node creation
I've hit this roadblock sooo many times. I try to update the underlying data, but I am unable to, because the library only returns read-only ref to current data. At the moment, only option to overcome this problem is to re-generate the node. This commit fixes that by adding methods (mutable and immuteable) for changing the underlying data.