Correct drawing and zoom when containing rect is not at (0, 0) (#115)
In my app I put a graph in a `Window` widget which can be freely moved
around. It appears that because nodes are always drawing at certain
screen positions, the actual graph just stays fixed relative to the app
frame, and the window widget sort of becomes a window into the graph.

Also, the zoom center is wrongly calculated because that is in fact
using relative coordinates to the parent rect, while the graph is always
in screen space.

This PR fixes both issues. When the parent widget is moved on screen,
the pan is adjusted so the graph moves with the window. Note that this
adjustment does not generate a panning event. I wasn't quite sure
whether that is desirable. I'll happily add it if you think that is
best.

Another approach I considered was not adjusting the pan, but instead
always draw everything relative to the containing rect. The reason I did
not opt for this approach was that it would likely break existing custom
draws. However, I do recon this is a generally better approach, but
perhaps that should be incorporated together with #108 when things are
going to break either way 🙂