Fix: replace 'instant' dependency with 'web-time' (#296)
# The Bug
While using this crate, I encounered an error with my project's wasm
build that went something like this:
```
Module not found: Error: Can't resolve 'env' in 'my-project/pkg'
```

# Finding the source
After a little searching, I found this [helpful
issue](https://github.com/orgs/community/discussions/16289#discussion-4062908)
on the wasm-bindgen repository. I used [this tool](
https://github.com/Liamolucko/find_wasm_import) (mentioned in the issue)
on my project to attempt to find the source of the error, and the source
was the instant crate.

# Solution
In all regards, web_time seems to be superior to instant. Instant even
mentions in [their
readme](https://github.com/sebcrozet/instant/blob/master/README.md) that
the crate is unmaintained and web_time should be considered as an
alternative.

---------

Co-authored-by: Dmitrii Samsonov <blitzar90@gmail.com>